requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=0.19.0

# Optional UI dependency
gradio>=4.0.0